requirements.txt
# CNN实验所需的Python库
# 安装命令: pip install -r requirements.txt

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理和机器学习
numpy>=1.21.0
scikit-learn>=1.0.0

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 图像处理
Pillow>=8.3.0

# 进度条显示
tqdm>=4.62.0

# 可选：Jupyter支持
# jupyter>=1.0.0
# ipykernel>=6.0.0